Finance Project using python and its libraries
The financial crisis of 2008 had a profound impact on the global economy, and the banking sector was one of the hardest-hit areas. In this project, we aim to perform an exploratory data analysis (EDA) of stock prices for major banks from January 1, 2006, to January 1, 2016. This period encompasses the years leading up to the crisis, the crisis itself, and the recovery phase that followed.